+++ core/src/com/angian/bobby/Carpet.java
@@ -1,0 +1,14 @@
+package com.angian.bobby;
+
+import com.badlogic.gdx.scenes.scene2d.Stage;
+
+public class Carpet extends Solid {
+    public Carpet(float x, float y, float width, float height, Stage s) {
+        super(x, y, width, height, s);
+    }
+/*    public Carpet(Stage s) {
+        //super(x, y, w, h, s);
+        //TODO
+    }
+*/
+}