+++ core/src/com/angian/bobby/Solid.java
@@ -1,0 +1,32 @@
+package com.angian.bobby;
+
+import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.graphics.g2d.Batch;
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
+import com.badlogic.gdx.scenes.scene2d.Stage;
+
+public class Solid extends BaseActor {
+    private static ShapeRenderer shapeRenderer = new ShapeRenderer();
+
+    public Solid(float x, float y, float width, float height, Stage s) {
+        super(x, y, s);
+        setSize(width, height);
+        setBoundaryRectangle();
+    }
+
+    /*
+    @Override
+    public void draw(Batch batch, float parentAlpha) {
+        batch.end();
+
+        shapeRenderer.setProjectionMatrix(batch.getProjectionMatrix());
+        shapeRenderer.setTransformMatrix(batch.getTransformMatrix());
+        shapeRenderer.begin(ShapeRenderer.ShapeType.Filled);
+        shapeRenderer.setColor(Color.YELLOW);
+        shapeRenderer.rect(getX(), getY(), getWidth(), getHeight());
+        shapeRenderer.end();
+
+        batch.begin();
+    }
+    */
+}