+++ core/src/com/angian/bobby/Styles.java
@@ -1,7 +1,6 @@
 package com.angian.bobby;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 //import com.badlogic.gdx.graphics.g2d.freetype.FreeTypeFontGenerator;
@@ -18,7 +17,7 @@ public class Styles {
     private static void initStylesBitmap() {
         BitmapFont customFont = new BitmapFont(Gdx.files.internal("digits_font.fnt"));
         customFont.getRegion().getTexture().setFilter(Texture.TextureFilter.Nearest, Texture.TextureFilter.Nearest);
-        customFont.getData().setScale(LevelLayout.SCALE_FACTOR);
+        customFont.getData().setScale(LevelConstants.SCALE_FACTOR);
 
         labelStyle = new Label.LabelStyle();
         labelStyle.font = customFont;