+++ core/src/com/angian/bobby/Swordsman.java
@@ -1,0 +1,11 @@
+package com.angian.bobby;
+
+import com.badlogic.gdx.scenes.scene2d.Stage;
+
+public class Swordsman extends BaseActor {
+    public Swordsman(Stage s) {
+        super(s);
+
+        //TODO
+    }
+}